Add support for hard reset through dedicated pin.
A hard reset (as opposed to a soft reset through AT+RST) has the advantage that it works even when the ESP8266's AT interface is hung for some reason. This can be used for reliably auto-resetting the ESP8266 by a watchdog.

To use this, wire a dedicated pin of the Arduino to the ESP8266's CH_PD line, and #defining WL_HARD_RESET_PIN somewhere in your code. Take the usual precautions regarding voltage (5V -> 3.3V), just as for the ESP8266's RX pin. The rationale for choosing the CH_PD instead of the RST line is explained by the analysis at https://hallard.me/esp8266-autoreset/. Basically he recommends CH_PD over RST for auto reset, because of problems with how the RST line is wired up internally.